+++ NEWS.md
@@ -8,6 +8,7 @@
 - The Windows installer now bundles the Tcllib library (#1136).
 - Fixed a bug where PSL directives in comments were parsed incorrectly
   when split over multiple lines (#1135).
+- Added support for PSL `prev()` built-in function (#1135).
 
 ## Version 1.15.0 - 2025-01-11
 - `--load` is now a global option and should be placed before the `-r`